+++ include/engine/runtime.h
@@ -23,6 +23,7 @@ class Runtime {
   private:
     void simTick(double dt);
     void dispatchAudioEvents();
+    bool loadAudioContent(const std::string& packSearchPath);
 
     EngineConfig config_;
     bool running_ {true};
@@ -30,10 +31,6 @@ class Runtime {
 
     InputSystem input_;
     AudioSystem audio_;
-    AudioSystem::SoundId hitSoundId_ {AudioSystem::kInvalidSound};
-    AudioSystem::SoundId grazeSoundId_ {AudioSystem::kInvalidSound};
-    AudioSystem::SoundId phaseWarnSoundId_ {AudioSystem::kInvalidSound};
-    AudioSystem::SoundId specialSoundId_ {AudioSystem::kInvalidSound};
     GameplaySession session_;
     RenderPipeline renderPipeline_;
 };